Fix #19

+++ src/ESP8266Ping.h
@@ -37,8 +37,8 @@ class PingClass {
   public:
     PingClass();
 
-    bool ping(IPAddress dest,   byte count = 5);
-    bool ping(const char* host, byte count = 5);
+    bool ping(IPAddress dest,   unsigned int count = 5);
+    bool ping(const char* host, unsigned int count = 5);
 
     int averageTime();
